requirements.txt
Django==1.10
django-countries==4.5
django-localflavor==1.5
django-material==1.0.0
django-registration-redux==1.6
